+++ packages/supermassive/bin/supermassive.js
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+
+(function () {
+  require("../lib/bin/supermassive").supermassive().parseAsync(process.argv);
+})();
