--- packages/supermassive/bin/typeDefsToImplicitResolvers.js
@@ -1,7 +1,0 @@
-#!/usr/bin/env node
-
-(function () {
-  require("@graphitation/supermassive/lib/bin/typeDefsToImplicitResolvers")
-    .typeDefsToImplicitResolvers()
-    .parseAsync(process.argv);
-})();
